requirements.txt
selenium
mysql-connector-python
requests
python-dotenv